requirements_SAVED.txt
Package             Version
------------------- -------
asgiref             3.8.1
backports.zoneinfo  0.2.1
beautifulsoup4      4.14.3
crispy-bootstrap5   2025.6
Django              4.2.26
django-bootstrap-v5 1.0.11
django-cleanup      9.0.0
django-crispy-forms 2.4
pillow              10.4.0
pip                 25.0.1
setuptools          75.3.2
soupsieve           2.7
sqlparse            0.5.4
typing_extensions   4.13.2
tzdata              2025.2
